\name{category.probs}
\alias{category.probs}
\title{Calculate the probabilities of each category given X}
\usage{
  category.probs(nn, X)
}
\arguments{
  \item{X}{Pattern from which we have to decide a category.
  It is a set of measurements represented by a
  p-dimensional vector}

  \item{nn}{An already trained probabilistic neural
  network}
}
\description{
  Calculate the probabilities of each category given X
}

