\name{fA}
\alias{fA}
\title{Sum the results of applying the kernel over all patterns}
\usage{
  fA(Xa, X, sigma)
}
\arguments{
  \item{X}{Pattern from which we have to decide a category.
  It is a set of measurements represented by a
  p-dimensional vector}

  \item{Xa}{One of the training patterns from category A}

  \item{sigma}{Smoothing parameter}
}
\description{
  Sum the results of applying the kernel over all patterns
}

