\name{guess.category}
\alias{guess.category}
\title{Decide the category of X}
\usage{
  guess.category(of.measurement, given.model)
}
\arguments{
  \item{of.measurement}{A vector describing a new
  observation}

  \item{given.model}{A probabilistic neural network already
  trained}
}
\description{
  Decide the category of X
}

