\name{guess.probabilities.of.each.category}
\alias{guess.probabilities.of.each.category}
\title{Predict the probabilities of each category given X}
\usage{
  guess.probabilities.of.each.category(nn, X)
}
\arguments{
  \item{nn}{An already trained probabilistic neural
  network}

  \item{X}{Pattern from which we have to decide a category.
  It is a set of measurements represented by a
  p-dimensional vector}
}
\description{
  Predict the probabilities of each category given X
}

