\name{patterns}
\alias{patterns}
\title{Apply kernel over all patterns from category A}
\usage{
  patterns(Xa, X, sigma)
}
\arguments{
  \item{Xa}{A vector describing a new observation}

  \item{X}{The training set of observation}

  \item{sigma}{The smooth parameter}
}
\description{
  Apply kernel over all patterns from category A
}

