\docType{data}
\name{skin}
\alias{skin}
\title{Skin segmentation}
\format{A data-frame with 245057 observations described by a color (R,G,B) and a binary variable (0/1) for the skin test}
\source{
  http://archive.ics.uci.edu/ml/datasets/Skin+Segmentation
}
\description{
  This source describes if a given colour (R,G,B) could be
  associated with the one of a skin.
}
\references{
  Rajen Bhatt, Abhinav Dhall, 'Skin Segmentation Dataset',
  UCI Machine Learning Repository
}
\keyword{datasets}

